Replaced old Overview Table with new Overview Cards
The old OverviewActivity was a replica of the schedule table on
the Genesis homepage - and it was ugly. This new design uses
a RecyclerView and CardViews, which are much newer and look a LOT
better. My goal for this app is to make it incorporate many
Material Design guidelines and patterns, and using Cards in the
OverviewActivity is certainly a step in the right direction.