Moved code from network.___Task to GenesisHTTP
Everything related to parsing and searching through
Genesis's HTML is now located in GenesisHTTP. Code was
taken from LoginTask, ClassTask, GradebookTask, and
OverviewTask. Because of this simplification,
GradebookTask is no longer needed and was merged into
OverviewTask.

The the animation library was updated in build.gradle as
well, which prevents the app from crashing due to
the library being incompatible with Google Play Services.