Added minor fixes and changes
One fix was that the app no longer crashes when you
click on a class with no grades. But now it shows the
first assignment for the year from any class instead.

Another change was that a new package (data) was added
and the login package was moved to there. I will add
more classes like LoginInfo soon, to organize the data
GenesisHTTP collects and make the code more neat.

Also, the URL for getting class pages was tweaked, and
it now gets assignments from MP2 instead of MP1. This
fix is temporary, and I will add code to automatically
select the current marking period soon.