OverviewActivity now refreshes its data
The OverviewActivity now gets new data every time
the activity is closed and opened again. The data it gets
is either from the app's cache (which was stored in the
last minute) or directly from the internet. Also, animation
is turned off for when the activity restarts, so it looks
smoother. However, when data is fetched from the
internet, there is a delay and a brief, but noticeable flash
when the old data is deleted and the new data is
inserted.