Added the ability to cycle between every marking period
On the OverviewActivity, the user now has the option to switch
between all four marking periods in the year. I don't know what
happens if a marking period that hasn't started yet is selected -
it might crash. Also, there are some problems with pressing the
menu button over and over again quickly, so that has to be fixed.
Also, when hitting the refresh button while a past marking period
is selected, the app ignores the selected marking period and
refreshes the current one, so that must be fixed as well.