Organized code and fixed a bug with reopening the app
After the user leaves the app and comes back to it, onStart() is
re-run. Previously, this would mean that the app would ignore
whichever marking period was selected in the menu and use the
default one, but that was fixed with this code. However, onStart()
does not rely on any form of cached data anymore, which would be
good to fix in the future.