Added a (rough) notification system and changed packeges
The notifications will alert the user when a new assignment
is posted. They are very crude at the moment, with a bad
icon (the refresh icon, when it should ideally be a white
G), and a very high chance of causing crashes. There are
many things to fix in this system; this commit was just to
implement the feature.

In order to find updates to the assignment lists, the app
needs to store the previous amounts of data in files. This
motivated a small package movement, where the LoginInfo
and LoginStore classes were moved to different locations.