Temporarily fixes a problem with the CardViews
A new version of the CardView library made them way
too big on the screen, and I have to quickly fix the
problem since I will be presenting the app as a project
tomorrow in school. I will put the library back to its
proper version later, but this temporary fix is good
enough for now.