Changed the alarm frequency from 15 mins to 30 mins
This app has been using lots of data in the background, so in an
attempt to reduce it I am lowering the frequency of notification
updates. I will be doing other things soon, such as increasing the
use of caching and maximizing the session cache time, so the data
usage will be lower.