Added a marking period cache
This is part of the movement towards reducing data usage. The app
now only needs to check for the current marking period once a
day, which will hopefully reduce the data usage by a good amount.

It's also worth noting that the Genesis webpages (just HTML alone)
are very large (~50KB per page). This means that the large data
usage by this app is unpreventable, and all I can do is minimize
its page visits. This is especially upsetting since the HTML
documents themselves have lots of whitespace, and could be reduced
by 50% if they were just compressed.