Added a License
For now, this app is under the MIT license. I might change it to the Apache license in the future though, since it forces any derivative work to be under the same license, whereas the MIT license does not have any restriction like this. Also, I should mention somewhere that JSoup is under the MIT license