Fixed the overview data collection function in GenesisHTTP
The Genesis website updated over the summer and no longer
places the course number next to the course name in its
Gradebook, so we have to obtain both in a slightly different way.
The name is taken by collecting the entire string (as opposed to
the second half, like before) and the number is taken by splitting
the HTML up and finding the exact position of the course number.