Fixed a small error with the MP cache being corrupt
Sometimes the phone would attempt to check for a new marking
period while there's no internet, so it would get "null" as the MP
and save it to the cache. It would then later use this value when
trying to gather other class information, and the Genesis website
would not respond correctly to "null" as a marking period.

This code simply returns false to the exists() function if the cache
is invalid, and contains a "null" marking period.